src/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.111.0
FlashRank==0.2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
pydantic==2.7.1
python-dotenv==1.0.0
qdrant-client==1.9.1
sentence-transformers==2.7.0
